Fix the bug regarding the game stopping at 4
Because I checked for `_sumOfPlayersScore === 4` a bug is caused when
the fourth game is a draw. The condition becomes true and the game stops.

To prevent this I moved the section to the end and checked for
`_sumOfPlayersScore === 5` instead. Even though there's a 3 seconds
pause before the winner is displayed, this is better.